requirements.txt
google-adk==0.5.0
google-genai==1.14.0
google-api-python-client
google-auth
google-auth-oauthlib
fastapi
uvicorn
python-dotenv
websockets
pydantic
crewai
langchain
langchain_community
langchain_core
langgraph
streamlit
deprecated
unstructured
mcp==1.9.1
